+++ ios/documentation/changelog.md
@@ -1,5 +1,8 @@
 # Change Log
 <pre>
+v1.2.2  [TIMODOPEN-269] Updated example to account for iOS 7 no longer supporting quickview of remote URLs.
+        [TIMODOPEN-269] Removed 'indexChanged' event
+
 v1.2.1	Building with 2.1.3.GA and open sourcing [MOD-1267]
 	
 v1.2	Added support for the "click" event